+++ pkg/client/guest_params.go
@@ -14,21 +14,21 @@ type PostDenyListBody struct {
 
 type GetGuestVisitsOptions struct {
 	site_id    string `name:"site_id"`
-	start_time *int64 `name:"start_time"`
-	end_time   *int64 `name:"end_time"`
+	start_time *int   `name:"start_time"`
+	end_time   *int   `name:"end_time"`
 	Page_token string `name:"page_token"`
-	Page_size  *int64 `name:"page_size"`
+	Page_size  *int   `name:"page_size"`
 }
 
 type GetGuestTypesOptions struct {
 	site_id string `name:"site_id"`
 	Cursor  string `name:"cursor"`
-	Limit   *int64 `name:"limit"`
+	Limit   *int   `name:"limit"`
 }
 
 type GetHostsOptions struct {
 	site_id string `name:"site_id"`
 	Email   string `name:"email"`
 	Cursor  string `name:"cursor"`
-	Limit   *int64 `name:"limit"`
+	Limit   *int   `name:"limit"`
 }
