+++ pkg/client/sensor_params.go
@@ -2,19 +2,19 @@ package client
 
 type GetSensorAlertsOptions struct {
 	device_ids []string `name:"device_ids"`
-	Start_time *int64   `name:"start_time"`
-	End_time   *int64   `name:"end_time"`
-	Page_size  *int64   `name:"page_size"`
+	Start_time *int     `name:"start_time"`
+	End_time   *int     `name:"end_time"`
+	Page_size  *int     `name:"page_size"`
 	Page_token string   `name:"page_token"`
 	Fields     []string `name:"fields"`
 }
 
 type GetSensorDataOptions struct {
 	device_id  string   `name:"device_id"`
-	Start_time *int64   `name:"start_time"`
-	End_time   *int64   `name:"end_time"`
+	Start_time *int     `name:"start_time"`
+	End_time   *int     `name:"end_time"`
 	Page_token string   `name:"page_token"`
-	Page_size  *int64   `name:"page_size"`
+	Page_size  *int     `name:"page_size"`
 	Fields     []string `name:"fields"`
 	Interval   string   `name:"interval"`
 }
